PR fixes and renamed PersonUtil -> LegalIdUtil
